refactor(map): restructure map feature into layered architecture (#16)
