+++ src/app/features/map/i18n/map-messages.ts
@@ -3,7 +3,6 @@ import { Injectable } from '@angular/core';
 @Injectable({
   providedIn: 'root',
 })
-
 export class MapMessagesService {
 
   readonly mapView = {
